factor out NavBarFlexContent
