+++ components/new-header/Appbar.css
@@ -1,8 +1,8 @@
 .appbarFlex {
   display: flex;
-  justify-content: space-between;
+  /* justify-content: space-between; */
   align-items: center;
-  flex-direction: row;
+  /* flex-direction: row; */
   width: 100%;
   height: 100%;
   padding: 0 20px;
@@ -16,8 +16,8 @@
 
 @media (max-width: 900px) {
   .appbarFlex {
-    justify-content: space-between;
-    align-items: center;
+    /* justify-content: space-between; */
+    /* align-items: center; */
     flex-direction: column;
     padding: 0;
   }
